Разделил зону ответственности классов. Разбил на более осмысленные блоки.
